// services/backend/src/enums/IndexStatus.enum.ts
export enum IndexStatusEnum {
  PROCESSING_REBALANCE = "PROCESSING_REBALANCE",
  PROCESSING_ETF_PRICE = "PROCESSING_ETF_PRICE",
  PROCESSING_ETF_FUNDING_REWARD = "PROCESSING_ETF_FUNDING_REWARD",
  SUCCESS_HISTORICAL = "SUCCESS_HISTORICAL",
  SUCCESS_REBALANCE = "SUCCESS_REBALANCE",
  SUCCESS_ETF_PRICE = "SUCCESS_ETF_PRICE",
  SUCCESS_ETF_FUNDING_REWARD = "SUCCESS_ETF_FUNDING_REWARD",
  PROCESSING_LIVE = "PROCESSING_LIVE",
  ERROR_NO_REBALANCE_DATA = "ERROR_NO_REBALANCE_DATA",
  ERROR_NO_CANDLES_DATA = "ERROR_NO_CANDLES_DATA",
  ERROR_NO_MARLETCAP_DATA = "ERROR_NO_MARLETCAP_DATA",
  ERROR_UNKNOWN = "ERROR_UNKNOWN",
}